use same Tomcat service setup for both CTP and CTS

+++ Artifacts/linux-continuous-testing-platform/linux_install_ctp.sh
@@ -40,9 +40,31 @@ tar xvzf apache-tomcat-$TOMCAT_VERSION.tar.gz
 mv apache-tomcat-$TOMCAT_VERSION /opt/tomcat
 echo "Tomcat 8 environment setup"
 export CATALINA_HOME=/opt/tomcat
-cp tomcat.sh /etc/init.d/
-chmod 755 /etc/init.d/tomcat.sh
-update-rc.d tomcat.sh defaults
+groupadd tomcat
+useradd -M -s /bin/nologin -g tomcat -d /opt/tomcat tomcat
+mkdir -p $CATALINA_HOME/conf/Catalina/localhost
+chgrp -R tomcat $CATALINA_HOME/conf
+chmod g+rwx $CATALINA_HOME/conf
+chmod g+r $CATALINA_HOME/conf/*
+chgrp tomcat $CATALINA_HOME
+chmod g+rwx $CATALINA_HOME
+chown -R tomcat $CATALINA_HOME/webapps/ $CATALINA_HOME/work/ $CATALINA_HOME/temp/ $CATALINA_HOME/logs/
+
+if [ -f /usr/sbin/update-rc.d ] ; then
+    echo "Using Update-rc to register Tomcat as a service"
+
+    cp tomcat.sh /etc/init.d/
+    chmod 755 /etc/init.d/tomcat.sh
+    update-rc.d tomcat.sh defaults
+
+elif [ -f /bin/systemctl ] ; then
+    echo "Using Systemd to register Tomcat as a service"
+
+    cp tomcat.service /etc/systemd/system/tomcat.service
+    systemctl daemon-reload
+    systemctl enable tomcat
+
+fi
 echo "cleanup"
 rm apache-tomcat-$TOMCAT_VERSION.tar.gz
 echo "Tomcat 8 installation finished"
@@ -76,10 +98,40 @@ unzip ctp_dist/em.war -d $CATALINA_HOME/webapps/em/
 cp license $CATALINA_HOME/webapps/em/
 cp database.properties $CATALINA_HOME/webapps/em/WEB-INF/classes/META-INF/spring/
 echo '<% response.sendRedirect("/em"); %>' >> /opt/tomcat/webapps/ROOT/index.jsp
+chown -R tomcat $CATALINA_HOME/webapps/em/
 
 echo "Remove temporary installation files"
 rm -rf ctp_dist
 rm parasoft_environment_manager_2.7.5.zip
 
 echo "Startup Tomcat"
-$CATALINA_HOME/bin/startup.sh
+
+if [ -f /usr/sbin/update-rc.d ] ; then
+    /etc/init.d/tomcat.sh start
+elif [ -f /bin/systemctl ] ; then
+    systemctl start tomcat
+else
+    su - tomcat -c $CATALINA_HOME/bin/startup.sh
+fi
+
+echo "Configure iptables to forward port 80 to 8080"
+
+iptables -A PREROUTING -t nat -i eth0 -p tcp --dport 80 -j REDIRECT --to-port 8080
+iptables -A PREROUTING -t nat -i eth0 -p tcp --dport 443 -j REDIRECT --to-port 8443
+
+if [ -f /usr/bin/apt ] ; then
+    echo "Using APT package manager"
+
+    export DEBIAN_FRONTEND="noninteractive"
+    echo iptables-persistent iptables-persistent/autosave_v4 boolean true | sudo debconf-set-selections
+    echo iptables-persistent iptables-persistent/autosave_v6 boolean true | sudo debconf-set-selections
+    apt-get -y install iptables-persistent
+
+elif [ -f /usr/bin/yum ] ; then 
+    echo "Using YUM package manager"
+
+    yum install -y iptables-services
+    systemctl enable iptables
+    service iptables save
+    systemctl start iptables
+fi
